clean code tasya
