+++ component/index.js
@@ -1,0 +1,1 @@
+export {default as TableSearch } from "./TableSearch/TableSearch"
